+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 3.0.1
 * Fixed subsetting of multivariate plots
+* The `force.symmetry` in `dtw_lb` is deprecated since it didn't serve any real purpose.
 
 ## Version 3.0.0
 * Bear in mind that the `DTW/SBD` algorithms (and hence the functions that depend on them) might give different results in installations with 32-bit architectures.